Fixing manual path to the prior model files for the posterior.
For the "max_logBF" calculation, if the individual is not
sequenced, set them to missing. I only want to calculate the max
BF based on the avilable data.

Delete the pre-compiled model files from the data/ directory - a
user may have to manually remove some ClinVar variants, so would
need to re-compute these anyway.