Fix bug in prior
Previously if the predictor had multiple values separated by a '&',
they were split and the max/min taken. But if there is one value,
it's saved as a float instead of a string, so no split needed.
Changing the code to check iff the predictor is a float or not
first.