Fix descendant table for malformed FAM
Where the parents do not preceed the children in the FAM file,
there will be errors with the descendant table. Mostly these are
edge cases, but I've modified the code to do the following:

- everyone is descended from themselves
- founders are not descended from anyone, their line is comlpete
- for all parents, their children are descended from them
- then, iterate through the pedigree. Identify individuals where
both parents are complete. Then these people are descended from
the people their parents are descended from.
- When no new entries have been set, there are no new ancestors
to account for.

This will be robust to any order of lines in the FAM file