BFDA changes
- Simulate the parameter terms, and allow for other Beta
distributions as a prior.
- Sample genotype vectors under the causal and neutral
models
- Included code to output graphs for the BFDA, including the
cutoffs to precision, TPE, and FPE values. The scale of these
plots is defined based on the observed logBF scales.
- Combine all info onto one plot rather than several separate
plots. Commented old code.
- Add BFDA to the main parser and add some command line options
to the general argument parser.

General update:
- If an individual has a phenotype code of -9, set them to control
- begin work on specifying a branch for BF calculations
- in the MRCA, remove founders from consideration if any
intervening relative between them and a carrier is a non-carrier.
- in the mas logBF, only set the genotypes of individuals in the
VCF file - everyone else is unknown.
- Some memory monitoring code has been added - for later deletion.